+++ src/notifications/models.py
@@ -10,6 +10,10 @@ class Notification(UUIDModel, TimeCreateModel):
     text = models.TextField(max_length=300)
     is_active = models.BooleanField(default=True)
     links = models.URLField()
+    account = models.ManyToManyField(
+        to="users.Account",
+        related_name="notifications"
+    )
 
     objects = models.Manager()
